fix: split oversized story files to comply with 500-line guard (#1226)
- TextareaAdvanced.stories.tsx (537→204): moved StateMatrix, FullMatrix to TextareaMatrix.stories.tsx
- InputVariants.stories.tsx (581→260): moved FocusTest, FormScenario, FullMatrix to InputMatrix.stories.tsx
- Heading.stories.tsx (513→418): moved FullMatrix to HeadingMatrix.stories.tsx
- Registered new story titles in surfaceRegistry.ts
- All aria-label attributes preserved
